requirements.txt
google-auth
google-auth-oauthlib
requests
python-dotenv
argparse
pytest
Pygments
markdown
colorama